Use Date rather than DateTime, change format
DateTime is likely overkill, and does not support addition/subtraction
out of the box.

The concern is "calendar date" (which Date implements), not "calendar
date with time" (which DateTime implements).

For simplicity, change the format from 'YYYYMMDD' to 'YYYY-MM-DD', which
is what the constructor for Date supports, and is ISO 8601 anyway.

Date objects are also more easily comparable than DateTime objects (as
can be seen in the test).