Clean up attributes
Use the private attribute twigil (eg. '$!', '@!') when declaring an
attribute that depends on another.

Get the value of an attribute via its private attribute twigil (eg.
'$!'), even if it's a public attribute (eg. '$.').

'cuts' was only public because 'self' was being used to access it:
'cuts' can just be private now.